requirements.txt
blinker==1.8.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
fonttools==4.57.0
gunicorn==23.0.0
idna==3.11
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==25.0
pillow==10.4.0
pycairo==1.26.1
pygobject==3.48.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
requests==2.32.4
scipy==1.10.1
six==1.17.0
tflite-runtime==2.14.0
urllib3==2.2.3
werkzeug==3.0.6
zipp==3.20.2
